+++ filament/src/PerViewUniforms.h
@@ -75,6 +75,7 @@ public:
     // maybe these should have their own UBO, they're needed only when GENERATING the shadowmaps
     void prepareShadowVSM(TextureHandle texture, VsmShadowOptions const& options) noexcept;
     void prepareShadowPCF(TextureHandle texture) noexcept;
+    void prepareShadowDPCF(TextureHandle texture) noexcept;
 
     // update local data into GPU UBO
     void commit(backend::DriverApi& driver) noexcept;